Refactoring starbBuild api
