+++ docs/swagger.json
@@ -72,31 +72,6 @@
                 "responses": {}
             }
         },
-        "/v1/images/delete/:id": {
-            "delete": {
-                "description": "update single data",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "meta Manager"
-                ],
-                "summary": "delete",
-                "parameters": [
-                    {
-                        "type": "string",
-                        "description": "The key for staticblock",
-                        "name": "id",
-                        "in": "path",
-                        "required": true
-                    }
-                ],
-                "responses": {}
-            }
-        },
         "/v1/images/get/{id}": {
             "get": {
                 "description": "get single one",
@@ -206,6 +181,22 @@
                         "required": true
                     }
                 ],
+                "responses": {}
+            }
+        },
+        "/v1/images/queryHistory/mine": {
+            "get": {
+                "description": "Query My History",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "meta Manager"
+                ],
+                "summary": "QueryMyHistory",
                 "responses": {}
             }
         },
