+++ util/export.go
@@ -2,7 +2,6 @@ package util
 
 import (
 	"bytes"
-	"fmt"
 	"strconv"
 )
 
@@ -26,8 +25,6 @@ type JsonData struct {
 
 //ExportData ExportData
 func ExportData(code int, title interface{}, data ...interface{}) *JsonData {
-
-	Log.Infoln(fmt.Sprintf("code:%d,title:%s,data:%v", code, title, data))
 	var resultData JsonData
 	resultData.Code = code
 	resultData.Title = title
