+++ util/log.go
@@ -46,9 +46,6 @@ func initLogger() *logrus.Logger {
 
 	Log.Out = io.MultiWriter(os.Stdout, src)
 
-	//info level
-	Log.SetLevel(logrus.InfoLevel)
-
 	Log.SetFormatter(&logrus.TextFormatter{
 		TimestampFormat: "2006-01-02 15:04:05",
 	})
